created a separate file to run loan tests loan-api.spec.ts
